backend/pyproject.toml
[tool.poetry]
name = "knowledge-window-backend"
version = "0.1.0"
description = "Knowledge Window Backend API"
authors = ["Devin <devin@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-multipart = "^0.0.17"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.2.1"
pypdf = "^5.1.0"
python-docx = "^1.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
